Rename abc -> base
